C-WCOW: Merge securitypolicy package for linux and windows
