+++ pkg/securitypolicy/securitypolicy_test.go
@@ -1,6 +1,3 @@
-//go:build linux
-// +build linux
-
 package securitypolicy
 
 import (
