+++ pkg/securitypolicy/securitypolicy_uvmpath_windows.go
@@ -1,0 +1,24 @@
+//go:build windows
+// +build windows
+
+package securitypolicy
+
+// This is being used by StandEnforcer and is a no-op for windows.
+// substituteUVMPath substitutes mount prefix to an appropriate path inside
+// UVM. At policy generation time, it's impossible to tell what the sandboxID
+// will be, so the prefix substitution needs to happen during runtime.
+func substituteUVMPath(sandboxID string, m mountInternal) mountInternal {
+	//no-op for windows
+	_ = sandboxID
+	return m
+}
+
+// SandboxMountsDir returns sandbox mounts directory inside UVM/host.
+func SandboxMountsDir(sandboxID string) string {
+	return ""
+}
+
+// HugePagesMountsDir returns hugepages mounts directory inside UVM.
+func HugePagesMountsDir(sandboxID string) string {
+	return ""
+}
